(#75) refactor: 팝오버 루트 팝오버로 변경
