+++ src/shared/components/ui/popover/index.ts
@@ -1,4 +1,4 @@
+export { default as Popover } from './Popover';
 export { default as PopoverContent } from './PopoverContent';
-export { default as PopoverRoot } from './PopoverRoot';
 export { default as PopoverTrigger } from './PopoverTrigger';
 export { usePopover } from './usePopover';
